Exercise 4.21 finished

+++ part4/blog/Controller/blog.js
@@ -55,6 +55,18 @@ blogRoutes.post('/', async (request, response) => {
 });
 
 blogRoutes.delete('/:id', async (request, response) => {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+        return response.status(401).json({ Error: 'Token invalid'})
+    }
+    const blog = await Blog.findById(request.params.id)
+    if (!blog){
+        return response.status(404).json({ Error: 'Blog not found'})
+    }
+    if (blog.user.toString() !== decodedToken.id.toString()){
+        return response.status(401).json({ Error: 'You cannot delete this blog'})
+    }
+    
     await Blog.findByIdAndDelete(request.params.id)
     response.status(204).end()
 })
@@ -72,3 +84,12 @@ blogRoutes.put('/:id', async (request, response) => {
 })
 
 module.exports = blogRoutes
+
+/*
+{
+    "title": "Exploring SQL queries",
+    "author": "Anonymous",
+    "url": "https://example.com/SQL",
+    "likes": 23,
+}
+*/
